+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 - Add read-only group to user table
 - Add read-only group to nofo index table and nofo edit page
 - Set NOFO group to user's group during import
+- Non-bloom users can only see NOFOs from their group
 
 ### Fixed